+++ REINVENT4Surfactants/scoring_functions/comp_pareto_gradient.py
@@ -311,11 +311,11 @@ class ParetoGradient:
         on disk (e.g. resuming a run), so correctness doesn't depend on
         whether this process has seen every step from the start."""
         if not Path(self.data_path).is_file() or os.stat(self.data_path).st_size == 0:
-            return pd.DataFrame(columns=["SurfTen", "pCMC"])
+            return pd.DataFrame(columns=["SurfTen", "pCMC"], dtype=float)
         previous_df = pd.read_csv(self.data_path)
         previous_df["SurfTen"] = 1 - previous_df["SurfTen"]
         previous_df["pCMC"] = 1 - previous_df["pCMC"]
-        return find_pareto_front(previous_df)[["SurfTen", "pCMC"]].reset_index(drop=True)
+        return find_pareto_front(previous_df)[["SurfTen", "pCMC"]].astype(float).reset_index(drop=True)
 
     def __call__(self, smiles: list[str]) -> ComponentResults:
         if self._front is None:
@@ -366,7 +366,7 @@ class ParetoGradient:
         # never needs reconsidering, so this is the only history the next
         # call needs (cheap: O((|front| + batch_size) log(...))).
         combined = pd.concat([self._front, smiles_df[["SurfTen", "pCMC"]]], ignore_index=True)
-        self._front = find_pareto_front(combined)[["SurfTen", "pCMC"]].reset_index(drop=True)
+        self._front = find_pareto_front(combined)[["SurfTen", "pCMC"]].astype(float).reset_index(drop=True)
 
         return ComponentResults([score])
